final changes made
