+++ _Readme.md
@@ -1,3 +1,17 @@
-A basic typing speed testing code that stores the players data in file..
-I will try  do CRUD operation in it..
-Rest coming soon................
+# Typing Game Console Application
+
+This console application is designed for a typing game that measures the typing speed of players and maintains a leaderboard.
+
+## Features
+
+- **New Game:** Players can enter their name and participate in a typing test to measure their speed.
+- **Top Five Typing Experts:** Displays the top five players based on their typing speed.
+- **CRUD Operations Implemented:**
+  - **C - Create:** Creating a file to input player's data.
+  - **R - Retrieve:** Retrieving player's data from the file.
+  - **U - Update:** Updating player data by adding it to the file and overwriting existing data.
+  - **D - Delete:** Deleting data from the file by overwriting the file, excluding the desired data.
+
+## Copyright & Contact
+
+© 2024 Just Another Human!! | [tupesh.com.np] | [tupeshghimire@gmail.com]
